Migrate upsert_like().
